+++ utils/gear-hooks/src/hooks/api/voucher/use-voucher-balance.ts
@@ -1,6 +1,8 @@
 import { HexString, generateVoucherId } from '@gear-js/api';
+import { useContext } from 'react';
 
 import { useBalance } from '../balance/use-balance';
+import { AccountContext } from 'context';
 
 function useVoucherBalance(programId: HexString | undefined, accountAddress: HexString | undefined) {
   const { balance, isBalanceReady } = useBalance(
@@ -10,4 +12,10 @@ function useVoucherBalance(programId: HexString | undefined, accountAddress: Hex
   return { voucherBalance: balance, isVoucherBalanceReady: isBalanceReady };
 }
 
-export { useVoucherBalance };
+function useAccountVoucherBalance(programId: HexString | undefined) {
+  const { account } = useContext(AccountContext);
+
+  return useVoucherBalance(programId, account?.decodedAddress);
+}
+
+export { useVoucherBalance, useAccountVoucherBalance };
