+++ utils/gear-hooks/src/hooks/index.ts
@@ -14,6 +14,8 @@ import {
   useIsAccountVoucherExists,
   useAccountVoucherBalance,
   useAccountVoucher,
+  useVouchers,
+  useAccountVouchers,
   useBalance,
   useBalanceFormat,
   useDeriveBalancesAll,
@@ -40,6 +42,8 @@ export {
   useIsAccountVoucherExists,
   useAccountVoucherBalance,
   useAccountVoucher,
+  useVouchers,
+  useAccountVouchers,
   useBalance,
   useBalanceFormat,
   useDeriveBalancesAll,
