+++ utils/gear-hooks/src/index.ts
@@ -14,6 +14,8 @@ import {
   useIsAccountVoucherExists,
   useAccountVoucherBalance,
   useAccountVoucher,
+  useVouchers,
+  useAccountVouchers,
   useBalance,
   useBalanceFormat,
   useDeriveBalancesAll,
@@ -69,6 +71,8 @@ export {
   useIsAccountVoucherExists,
   useAccountVoucherBalance,
   useAccountVoucher,
+  useVouchers,
+  useAccountVouchers,
   useBalance,
   useBalanceFormat,
   useDeriveBalancesAll,
